most bugs fixed up, scene loads and connects properly
