+++ packages/client/src/components/World/LoadEngineWithScene.tsx
@@ -1,12 +1,11 @@
-import React, { useEffect, useRef } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import { useHistory } from 'react-router'
-import { initNetwork, loadLocation } from './LocationLoadHelper'
+import { initClient, initEngine, loadLocation } from './LocationLoadHelper'
 import { EngineActions, useEngineState } from '@xrengine/engine/src/ecs/classes/EngineService'
 import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
 import { SceneService, useSceneState } from '@xrengine/client-core/src/world/services/SceneService'
 import { Downgraded } from '@hookstate/core'
 import { InstanceConnectionService } from '@xrengine/client-core/src/common/services/InstanceConnectionService'
-import { MediaStreamService } from '@xrengine/client-core/src/media/services/MediaStreamService'
 import {
   LocationAction,
   LocationService,
@@ -15,34 +14,11 @@ import {
 import { useDispatch } from '@xrengine/client-core/src/store'
 import { leave } from '@xrengine/client-core/src/transports/SocketWebRTCClientFunctions'
 import { getWorldTransport } from '@xrengine/client-core/src/transports/SocketWebRTCClientTransport'
-import { useWorld } from '@xrengine/engine/src/ecs/functions/SystemHooks'
-import { NetworkWorldAction } from '@xrengine/engine/src/networking/functions/NetworkWorldAction'
-import { updateNearbyAvatars } from '@xrengine/engine/src/networking/systems/MediaStreamSystem'
 import { useProjectState } from '@xrengine/client-core/src/common/services/ProjectService'
-import {
-  configureClient,
-  createEngine,
-  initializeBrowser,
-  initializeCoreSystems,
-  initializeRealtimeSystems,
-  initializeSceneSystems
-} from '@xrengine/engine/src/initializeEngine'
 import { teleportToScene } from '@xrengine/engine/src/scene/functions/teleportToScene'
-import matches from 'ts-matches'
-import { SystemModuleType } from '@xrengine/engine/src/ecs/functions/SystemFunctions'
+import { Network } from '@xrengine/engine/src/networking/classes/Network'
 
 const engineRendererCanvasId = 'engine-renderer-canvas'
-
-const injectedSystems: SystemModuleType<any>[] = [
-  {
-    type: 'FIXED',
-    systemModulePromise: import('@xrengine/client-core/src/systems/XRUILoadingSystem')
-  },
-  {
-    type: 'FIXED',
-    systemModulePromise: import('@xrengine/client-core/src/systems/AvatarUISystem')
-  }
-]
 
 const canvasStyle = {
   zIndex: 0,
@@ -60,15 +36,6 @@ interface Props {
   setLoadingItemCount?: any
 }
 
-const initClient = async () => {
-  createEngine()
-  initializeBrowser()
-  configureClient()
-  await initializeCoreSystems(injectedSystems)
-  await initializeSceneSystems()
-  await initializeRealtimeSystems()
-}
-
 export const LoadEngineWithScene = (props: Props) => {
   const locationState = useLocationState()
   const history = useHistory()
@@ -76,12 +43,11 @@ export const LoadEngineWithScene = (props: Props) => {
   const engineState = useEngineState()
   const sceneState = useSceneState()
   const projectState = useProjectState()
+  const [clientInitialized, setClientInitialized] = useState(false)
+  const [clientReady, setClientReady] = useState(false)
 
-  /**
-   * Fetch projects so we know what we need to load into the engine
-   */
   useEffect(() => {
-    initNetwork()
+    initEngine()
   }, [])
 
   /**
@@ -89,22 +55,11 @@ export const LoadEngineWithScene = (props: Props) => {
    */
   useEffect(() => {
     // We assume that the number of projects will always be greater than 0 as the default project is assumed un-deletable
-    if (!Engine.isInitialized && !Engine.isLoading && projectState.projects.value.length > 0) {
+    if (!clientInitialized && Engine.isInitialized && projectState.projects.value.length > 0) {
+      setClientInitialized(true)
       const projects = projectState.projects.value.map((project) => project.name)
       initClient().then(() => {
-        useWorld().receptors.push((action) => {
-          matches(action)
-            .when(NetworkWorldAction.createClient.matches, () => {
-              console.log('CLIENT RECEPTORS', action)
-              updateNearbyAvatars()
-              MediaStreamService.triggerUpdateNearbyLayerUsers()
-            })
-            .when(NetworkWorldAction.destroyClient.matches, () => {
-              console.log('CLIENT RECEPTORS', action)
-              updateNearbyAvatars()
-              MediaStreamService.triggerUpdateNearbyLayerUsers()
-            })
-        })
+        setClientReady(true)
       })
     }
   }, [projectState.projects.value])
@@ -123,12 +78,12 @@ export const LoadEngineWithScene = (props: Props) => {
    * Once we have the scene data, initialise the engine
    */
   useEffect(() => {
-    if (locationState.currentLocation.location.sceneId.value && sceneState.currentScene.value) {
+    if (clientReady && locationState.currentLocation.location.sceneId.value && sceneState.currentScene.value) {
       dispatch(EngineActions.setTeleporting(null!))
       const [project] = locationState.currentLocation.location.sceneId.value.split('/')
       loadLocation(project, sceneState.currentScene.scene.attach(Downgraded).value!)
     }
-  }, [locationState.currentLocation?.location?.sceneId?.value, sceneState.currentScene?.scene?.value])
+  }, [clientReady, locationState.currentLocation?.location?.sceneId?.value, sceneState.currentScene?.scene?.value])
 
   const canTeleport = useRef(true)
   useEffect(() => {
