+++ packages/engine/src/scene/functions/SceneLoading.ts
@@ -64,7 +64,6 @@ export const loadSceneFromJSON = async (sceneData: SceneJson, world = useWorld()
 
   await Promise.all(Engine.sceneLoadPromises)
 
-  await delay(1000)
   Engine.sceneLoaded = true
 
   // Configure CSM
